+++ packages/logger/test/util.ts
@@ -72,5 +72,6 @@ export const finishLogger = async (logger) => {
 }
 
 export const getCurrentDateString = () => {
-  return new Date().toISOString().split('T')[0];
+  const d = new Date();
+  return `${d.getFullYear()}-${(d.getMonth() + 1).toString().padStart(2, '0')}-${(d.getDate()).toString().padStart(2, '0')}`
 };